+++ CHANGELOG.md
@@ -15,6 +15,7 @@ For a full diff see [`5.15.1...main`][5.15.1...main].
 ### Removed
 
 - Removed `$name` field from `Config\Ruleset\AbstractRuleSet` ([#865]), by [@localheinz]
+- Removed `$targetPhpVersion` field and `targetPhpVersion()` method from `Config\Ruleset\AbstractRuleSet` ([#866]), by [@localheinz]
 
 ## [`5.15.1`][5.15.1]
 
@@ -1152,6 +1153,7 @@ For a full diff see [`d899e77...1.0.0`][d899e77...1.0.0].
 [#856]: https://github.com/ergebnis/php-cs-fixer-config/pull/856
 [#857]: https://github.com/ergebnis/php-cs-fixer-config/pull/857
 [#864]: https://github.com/ergebnis/php-cs-fixer-config/pull/864
+[#866]: https://github.com/ergebnis/php-cs-fixer-config/pull/866
 
 [@dependabot]: https://github.com/apps/dependabot
 [@linuxjuggler]: https://github.com/linuxjuggler